+++ packages/mobile/system/createThemeConfig.ts
@@ -7,18 +7,21 @@ import { CreateThemeConfigParams } from './ThemeConfig';
 
 const getCacheKey = ({
   name,
+  spectrum,
   parentThemeConfig,
-}: Pick<CreateThemeConfigParams, 'name' | 'parentThemeConfig'>) => {
+}: Pick<CreateThemeConfigParams, 'name' | 'parentThemeConfig' | 'spectrum'>) => {
   const parent = parentThemeConfig ? `${parentThemeConfig.name}-` : '';
-  return `${parent}${name}`;
+  const spectrumKey = spectrum ? `-${spectrum}` : '';
+  return `${parent}${name}${spectrumKey}`;
 };
 
 export const createThemeConfig = memoize(function createThemeConfig({
   palette,
+  spectrum,
   parentThemeConfig,
   name: nameProp,
 }: CreateThemeConfigParams) {
-  const name = getCacheKey({ name: nameProp, parentThemeConfig });
+  const name = getCacheKey({ name: nameProp, parentThemeConfig, spectrum });
   return {
     name,
     light: createThemeConfigForSpectrum({
@@ -41,6 +44,7 @@ export function useFallbackThemeConfig() {
   const spectrum = useSpectrum();
   return createThemeConfig({
     palette: spectrum === 'dark' ? darkDefaultPalette : defaultPalette,
+    spectrum,
     name: 'default',
   });
 }